Adding Descriptions to Links
I really like how the links are organized here in one neat place. While doing research I kept coming back to this page as a reference to see if I was duplicating Trasks's research. It's hard to tell what the links are though just by glancing at them so I thought labeling might help, esp if someone wants to find something specific or just read 2-3 items that are of particular interest. If that's helpful I'll go ahead and do the rest of the links too :)