Federated Learning with Differential Privacy on Remote Machine
When effectively implementing Federated Learning in real world applications, it is important to build an architeture where data in a remote server could be accessed in order to train and test the model.

In the previous examples this is implemented using virtual clients. This example attempts to implement it over websockets.