Fix test
This test was broken because it relied on the remote worker having logging engaged.
However, somehow I forgot to turn on logging before. It could be that since the
worker was global, if tests are run in a correct order they would pass. At any rate,
I figured out the issue. I also had to remove one assertion from
test/message/test_message.py::test_is_none_message because our serializer raises
an error on it (which is a valid error). Basically, it was checking to see if the
msg typ was correct after the tensor had been deleted, but we cannot deserialize
the msg whn inside the message it neeeds to deserialize a tensor which no longer
exists. Not a big deal, it wasn't an important assertion.

The bug regarding the worker (bob) not having log_msgs == True seems to be affecting
several tests. I'll be fixing those next