Add documentation to Message class
I added some inline documentation/comments to the Message class. I left one TODO
which I could have done myself, but it wouldn't get used in the current state
of the codebase, and it's a nice self-contained project which could be done
as a first project for someone learning the codebase.