Add TODO
As an additional option (see last commit), I added an alterative for the project. It's unclear
to me at the moment whether it's a good idea for the Message type to have a default detailer.
It's not harming anything at the moment, but I also don't think it's getting used.