Custom message types (#2450)
* added new custom message types, but no change in functionality yet

* Added unit tests for CommandMessage, ObjectMessage, and ObjectRequsetMessage

* added tests for GetShapeMessage and ForceObjectDeleteMessage

* balck

* formatting

* __getitem__ should return none if the item doesn't exist

* bugfix with todo for pointer.is_none

* test isnone message

* black

* formatting

* added searchmessage with test

* black

* todos

* black

* added faster detailer and simplifier for CommandMessage

* black

* fixed comments

* fixed init

* black

* removed unnecessary comment

* cleanup request from bobby

* added _get_msg utility function

* added compress and decompress to serde init

* Add type checking for messages

In the PR, bobby made a recommendation to make sure th right messages are sent.
I actually found a bug when doing this which got sorted as well.

* run black

* Fix a typo

* Add documentation to message.py

* Fix test

This test was broken because it relied on the remote worker having logging engaged.
However, somehow I forgot to turn on logging before. It could be that since the
worker was global, if tests are run in a correct order they would pass. At any rate,
I figured out the issue. I also had to remove one assertion from
test/message/test_message.py::test_is_none_message because our serializer raises
an error on it (which is a valid error). Basically, it was checking to see if the
msg typ was correct after the tensor had been deleted, but we cannot deserialize
the msg whn inside the message it neeeds to deserialize a tensor which no longer
exists. Not a big deal, it wasn't an important assertion.

The bug regarding the worker (bob) not having log_msgs == True seems to be affecting
several tests. I'll be fixing those next

* Fix test_force_object_delete_message

This test was broken in the same way as the previous commit. bob did not have
logging engaged, so we couldn't use the message log in our assertions

* Fix test_get_shape_message

This test was broken in the same way as the previous commit. bob did not have
logging engaged, so we couldn't use the message log in our assertions

* Fix test_obj_req_message

This test was broken in the same way as the previous commit. bob did not have
logging engaged, so we couldn't use the message log in our assertions

* Fix test_obj_message

This test was broken in the same way as the previous commit. bob did not have
logging engaged, so we couldn't use the message log in our assertions

* Fix test_cmd_message

This test was broken in the same way as the previous commit. bob did not have
logging engaged, so we couldn't use the message log in our assertions

* Run black

* Run black

I forgot to run black on syft as well in the previous commit

* Add documentation to Message class

I added some inline documentation/comments to the Message class. I left one TODO
which I could have done myself, but it wouldn't get used in the current state
of the codebase, and it's a nice self-contained project which could be done
as a first project for someone learning the codebase.

* Add TODO

As an additional option (see last commit), I added an alterative for the project. It's unclear
to me at the moment whether it's a good idea for the Message type to have a default detailer.
It's not harming anything at the moment, but I also don't think it's getting used.

* Add inline documentation for CommandMessage

Details for this commit can be viewed in the documentation that I wrote.

* Fix error in docs

While working on the last commit I noticed an error in Message's inline documentation so I fixed it

* Typo in CommandMessage docs

* Add documentation for ObjectMessage

* Add documentation for ObjectRequestMessage

* Add documentation for IsNoneMessage

Deetailsf or this commit can be viewd by reading th documentation added

* Change messaging import strategy in serde.py

In the PR for this work, @midokura-silvia requested that imports be done differently.
https://github.com/OpenMined/PySyft/pull/2450#discussion_r312815181
This follows inline with her recommendation.

* Access Message.contents as property

@midokura-silvia had a reasonable request to always save contents in Message instantiations. However, this is less efficient and breaks when extensions of the message have their own ways for storing their contents (other than a generic tuple). Explanation for why these are important has been added to the documentation in this commit.

I chose to alter the .contents attribute of Messagese to be a property which referencse a private attribute ._contents. This satisfies
@midokura-silvia's request to keep the number of attributes fixed while also allowing for the efficiency gains
of messages with more expressive content storage.

* Run 'black syft'

* Fix typo in several todos 'detalier' -> 'detailer'

* Finish IsNoneMessage documentation

I committed a bit early on that one

* Change tensor_tuple to msg_tuple

* Document GetShapeMessage type

same as previous commit but for this messag etype

* Document ForceObjectDeleteMessage type

same as previous commit but for this messag etype

* Document SearchMessage type

same as previous commit but for this messag etype

* Add Github issues for each TODO

As requested from https://github.com/OpenMined/PySyft/pull/2450#discussion_r314321089
I have added a github issue for each TODO inside of syft/messaging/message.py

* Run 'black syft'