Add 'syft/messaging/promise.py'
This will be the file in which the new Promise type will be created. This will allow
a worker to tell another worker that in the future it will have a tensor with a certain
id. Other workers can then add operations to that promise such that when the promise is
kept the other operations will be automatically triggered