Move default plans argument to initializer body
There's a very strange garbage collecion bug when you put set() in the
default value part of an argument in __init__. When you re-initialize
a variable with the same name it gives it the same object as the last
plan had. Python should fix this bug.

At any rate, the fix for us is simple. Set plans=None by default and
just reset it to set() if 'plan is None'