Fix multi-inheritance for PromisTensor
It was really tricky to get PromiseTensor to properly call in the initializers
for AbstractTensor and Promise but I did get it to work. I'm not 100% sure it's
the conventional way. https://stackoverflow.com/questions/9575409/calling-parent-class-init-with-multiple-inheritance-whats-the-right-way
was helpful and a few other resources, but I ultimately had to try a combination
of multiple strateegies together. I *think* it's ok but I'm not totally sure