Fix typing issue in wrappers
This commit was a little bit painful. Basically, I wanted PromiseTensors to
automatically transform into the tensors they're supposed to be, even if they're
wrapped. At the last commit thi kindof worked, but it ended up with tensors
like Wrapper>tensor([1,2,3,4,5]) so I needed to instead move the data
into the wrapper. But this presented a problem because our wrapper types don't
match the data they wrap. So, I had to create a new .torch_tensor() method which
would try to sort out what type wrappers should be. I'll be surprised if this
functionality is perfect at this point but it works for this part of the codebase.