Generate __add__ dynamically
Previous to this commit I had hard-coded th __add__ function for TensorPlan,
but since I do not want to do this for all functions, I modified the __add__
code to generate it automatically. In the next commit, I'll move this code to
TorchHook and try to do it for all methods.