Find functions to hook using user supplied module name
Tensorflow uses swig to generate APIs and thus their module structure is
quite different than their public api.

in `_get_hooked_func` we would construct the command name to executed like so

```
cmd_name = f"{attr.__module__}.{attr.__name__}"
```

In torch, this was never a problem.  However, with Tensorflow, the attribute
module would often times not line up with the public API.  This uses the user supplied
API namespace for hooking so we can be more confident the API will be there.