Add experimental support for mis-matching args
One thing we want to be able to do is to add a tensor of scalars to a tensor of
encrypted valus and vise versa. This was very tricky to do and we need to
look closer to make sure I didn't add any security issues but I think
it looks right.