PaillierTensor and Arg Type Matching (experimental) (#2740)
* Cleanup directory names

With syft.frameworks.torch, the directory names have becomee disorganized
and don't seem to follow any set standard. Also, some are simply too long
and cumbersome to be typing out all the time (such as differential_privacy).

I have replaced long names with their shorthand options, federated leearning
to simply fl, differential_privacy to dp. I also changed crypto to mpc
since all of the algorithms contained within the folder exclusively
related to secure multi-party computation. If and when we implement
a wider varitey of crypto algorithms, we can reconsider this name. However,
at present, I would like for us to use a more descriptive directory name.

* Revert changed notbooks

I testeed all the notebooks to ensure that the work with the
renamd directories but don't wish to actually change them

* Revert changed notbooks

I testeed all the notebooks to ensure that the work with the
renamd directories but don't wish to actually change them

* Revert setup.py change

* Revert docs change

* Revert docs change

* Revert docs change

* Init boilerplate example tensor

I am beginning this project by copy-pasting LoggingTensor as a boilerplate
example of how to create a custom tensor type in PySyft. I then changed
all instances of LoggingTensor to PaillierTensor

* Init experimental notebook

I'll be doing my experimenting and development/testing in this notebook

* Add PaillierTensor to hook_args

In this commit, I added PaillierTensor to the hook_args
configuration file

* Add support for encrypt, decrypt, and __add__

I added support for basic encryption, decryption, and addition to PaillierTensor.
Data within the tensor is stored as a numpy array of phe scalars, where phe
is the python package for paillier homomorphic encryption. This will allow
us to use many of the desirable numpy operations as needed

* Remove unused boilerplate

* Parallelize homomorphic encryption and decryption

Since the encryption and decryption step was quite slow, I parallelizd it
using the multiprocessing library. I'm pretty sure I'll be able to do this
for all of the paillier operations, greatly increasing the speed of the lib.

* Move thread pool to syft.

I don't want to have to re-initialize the thread pool over and over, so I'm
adding it as a global syft variable

* Hook args automatically

Since PyTorch has a lot of similar methods to Numpy, by defaeult se can use
the normal hooking logic for most methods

* Add experimental support for mis-matching args

One thing we want to be able to do is to add a tensor of scalars to a tensor of
encrypted valus and vise versa. This was very tricky to do and we need to
look closer to make sure I didn't add any security issues but I think
it looks right.

* Add comment

* Remove unused code

* Add support for encrypted matrix multiplication

* Fix bug in arithmetic

* Fix bug in matmul

* Fix bug in test

* Remove verbose comments

* Run black

* Add phe to requirements.txt