Cleanup directory names (#2739)
* Cleanup directory names

With syft.frameworks.torch, the directory names have becomee disorganized
and don't seem to follow any set standard. Also, some are simply too long
and cumbersome to be typing out all the time (such as differential_privacy).

I have replaced long names with their shorthand options, federated leearning
to simply fl, differential_privacy to dp. I also changed crypto to mpc
since all of the algorithms contained within the folder exclusively
related to secure multi-party computation. If and when we implement
a wider varitey of crypto algorithms, we can reconsider this name. However,
at present, I would like for us to use a more descriptive directory name.

* Revert changed notbooks

I testeed all the notebooks to ensure that the work with the
renamd directories but don't wish to actually change them

* Revert changed notbooks

I testeed all the notebooks to ensure that the work with the
renamd directories but don't wish to actually change them

* Revert setup.py change

* Revert docs change

* Revert docs change

* Revert docs change