Translation to Spanish of the tutorial Part 02
feat: Translation to Spanish of the Part 02 - Intro to Federated Learning.

This comes as a response to the initiative to translate all the
PySyft tutorials, as it is desirable to learn PySyft in our native
language.

As requested, every part of the tutorial is done via pull request
separately to:

 - Track better the changes in every document.

 - Avoid double work.

Resolves: #2775
See also: #2782