Split `message` attr of `Operation` into separate attrs
In theory, this class is generic and can accept any kind of data
structure as a message, but in practice it's only used to capture ops
that have a certain calling structure. We can let go of the genericness
to get a class that's a bit easier to work with. This is a first step in
that direction.