Change `Operation` constructor to take named args (instead of tuple) (#2850)
* Split `message` attr of `Operation` into separate attrs

In theory, this class is generic and can accept any kind of data
structure as a message, but in practice it's only used to capture ops
that have a certain calling structure. We can let go of the genericness
to get a class that's a bit easier to work with. This is a first step in
that direction.

* Change `Operation` constructor to take named args (instead of tuple)

* Standardize `Operation` field names

* Update the `Operation` serde tests for the new field names

* Fix one additional call to `Operation` constructor in `BaseWorker`