Only run the tutorial translation tests when the tutorials change
The normal tutorial notebook tests still run on every build, which
should catch the cases where changes to the code break the notebooks.
When the main tutorial notebooks are changed, that will trigger the
action that runs the notebook translation tests, which should surface
the same failures in the whichever translated versions of the tutorials
are affected.