* Translation to Spanish of the tutorial Part 07
feat: Translation to Spanish of Part 07 - Federated Learning with Federated Dataset.

This comes as a response to the initiative to translate all the
PySyft tutorials, as it is desirable to learn PySyft in our native
language.

As requested, every part of the tutorial is done via pull request
separately to:

 - Track better the changes in every document.

 - Avoid double work.

Resolves: https://github.com/OpenMined/PySyft/issues/2775
See also: https://github.com/OpenMined/PySyft/issues/2782