Split the (English) tutorial tests into a separate Github Action (#3075)
* Split the (English) tutorial tests into a separate Github Action

The tutorial tests take about 15 minutes to run, which is approximately
equal to the running time of all the rest of the tests. Running them in
parallel should speed up the PR checks.

* Exclude the isolated file system test fixture from coverage checks

It's only used by the notebook tests, which are run separately as of
this pull request.

* Run notebook tests on all PRs except translation tests

Before this change, they would only run on PRs that changed notebooks.