Init generic ObjectConstructor class with Torch instance
In this commit, I created the generic ObjectConstructor class which allows us to
implement functionality before and after the constructor for any arbitrary
object type which will be very convenient for extending very differenet frameworks
using (more or less) the same approach.

In this commit, I also created two subclassese of ObjectConstructor necessary to
add functionality before or after the constructors of torch tensors and torch
paramters.