Update path triggers for automated test workflows (#3328)
* Update path triggers for automated test workflows

The test suite includes some flaky tests, which are hampering our ability
to merge translation PRs that don't touch the main code of the library.
This updates the paths that trigger the test suite so that it only runs
when `.py` requirements files are changed. The tutorials tests will only
run when `.py`, `.ipynb`, or requirements files are changed. The
translation tests will only run when there are changes to the translation
notebooks.

* Update the names of the workflow jobs