Fix missing `device` param in `test_federated_client.py::test_fit`
This test was failing unexpectedly in the automated test suite, which
appears to be due to randomization sometimes leading the `loss` to
increase instead of decrease. In that circumstance, the test would try
to train the model further until the loss decreased, but the line that
trains the model again was missing the `device` param, causing the test
to fail intermittently.