Remove test_spinup_time
Time based tests are inherently flaky based off of hardware. The original idea
of this test was to make sure that virtual workers were created quickly since
we were concerned about virtual workers taking too long to be created (since
data had to be copied) and wanted to make sure that we didn't regress to a point
where they were so slow that they caused timeout errors. The timings and amount
of data used to test were arbitrary. By removing this test we can reduce test
flakiness and add allow for more freedom to add something similar back when the
requirements are better understood.