Register the `translation` mark with PyTest to suppress warnings (#3618)
This mark is used in the test suite to designate tests that should only
run during the translation tests and not as part of the regular notebook
tests. It works fine now, but results in warnings from PyTest at the end
of the test suite output about an unregistered mark. This change makes
the warnings go away, but otherwise doesn't change anything about the
functionality of the test suite.