Create `AbstractPointer` and `AbstractSendable.create_pointer()` method
In the future, this will allow us to require that all sendable objects
are able to create a pointer to themselves. We can't do that yet,
because many of the custom tensor types have been relying on the tensor
wrappers to provide pointers. We'll need to both remove the wrappers and
make it possible for all the custom tensor types to create pointers to
themselves.