new strategy for storing rank to id translation tables
Each CrypTen computation is identified using a cid, it might be
different in different workers. Translation tables are indexed using
CIDs. Global table and the CID of the running computation are stored
under syft.frameworks.crypten, we assume that parties are separate
processes and that CID is overwritten after process creation, which
means that even if parties are running in the same machine, they won't
manipulate the global CID concurrently.