Refactor codebase
In this commit, I moved the modules which I expect to eventually drop down
to a lower-level language such as C++ or Rust into their own module called
'core'. This will help us have a separation of concerns. In theory, nothing
in 'core' should import anything from outside of core.