Further simplify serialization (#3966)
* Modify notebook

* greatly simplify serde logic

* Remove extraneous print

* Recomment out Duet

* Add inline documentation to MetaSerializable

* Add serialization convenience methods to torch objects

* Remove extraneous print

* Simplify hex and binary serialization
While I think it's a fine idea to make the hex serde more compressed, it did
so at the expense of hex working automatically. In order to deserialize,
you have to provide a specific type to deserialize to. We have no way of
knowing this, so the samller hex serialization cannot be used. It is better
to have a larger binary that can be deserialized anytime than to have
a smaller one which cannot be deserialized without already knowing
what is inside.

* Remove schema_type from deserialize()