Working towards serialization of SaveObjectAction (#3968)
* Modify notebook

* greatly simplify serde logic

* Remove extraneous print

* Recomment out Duet

* Add inline documentation to MetaSerializable

* Add serialization convenience methods to torch objects

* Remove extraneous print

* Simplify hex and binary serialization
While I think it's a fine idea to make the hex serde more compressed, it did
so at the expense of hex working automatically. In order to deserialize,
you have to provide a specific type to deserialize to. We have no way of
knowing this, so the samller hex serialization cannot be used. It is better
to have a larger binary that can be deserialized anytime than to have
a smaller one which cannot be deserialized without already knowing
what is inside.

* Remove schema_type from deserialize()

* Restore duet code

* Rename duet experiment notebook

* Add unfinished serialization logic for SaveObjectAction

* Remove extraneous functionality from AbstractNode and AbstractNodeClient

* Add Unspecified and All locations back

* Init SpecificLocation object

* Begin refactor based on simplification of codebase

This is really quite challenging but I'm most of the way there.
It's quite late here so I need to checkpoint and continue
later. I'm working through the Early Dev.ipynb

* Finish initial refactor sans tests

* Fix import ins torable tests

* All tests pass

* Init high level node message passing tests

* ADd torch tensor tests

* Rename test file so that tests run

* Run precommit script

* StorableOBject has id not key

* StorableOBject has id not key

* Fix code based on pre-commit hook warnings

* Comment out dangerous pickle import to please CI

* Increase code coverage requirements to new level