Begin code for the day
So I'm starting my coding today working on https://github.com/OpenMined/PySyft/pull/3978.
The premise for this project is to get Duet working far enough
so that I can call .send() on torch tensors and send them to
a duet server, and then call .get() on the resulting pointer
and get the tensor back.

So to get started, I'm working in the examples/experimental/trask/Duet Experiments.ipynb
notebook. I start by trying to send a tensor to a duet server and
the main complaint is

TypeError: missing a required argument: 'address'

I remember that this is because I forgot to add address to the
new protobuf file I created for this SaveObjectMessage. And
now in order to do that I need to serialize Address.

but in order to serialize Address, I need to first serialize
SpecificLocation because the address I'm trying to use here
contains a series of SpecificLocation objects.

So that's where my project starts today. Serializing
SpecificLocation as mentioned in the PR I'm working on.'