Improve SpecificLocation.__repr__ by improving ObjectWithId.__repr__
The specific location REPR was returnning ObjectWithID by default because
it inherits from it. However, this was a bit misleading. At first I thought
I'd just overload the method in SpecificLocation but I found a better way
which will make all subclasses from ObjectWithID print a more informative
.__repr__() so I changed that instead.

I also had to fix a test which broke because of this improvement.