Improve broken Address().__repr__()
It was printing strangely as an artifcat of the previous address refactoring

I've also noticed that when address has lots of UIDs that it becomes
really long and difficult to read. We need a short representation
of the UID for when its embedded. I think i'll add that while I'm here.