[WIP] Send/Get object to/from Duet server. (#3978)
* Increase code coverage requirements to 81%

* Begin code for the day

So I'm starting my coding today working on https://github.com/OpenMined/PySyft/pull/3978.
The premise for this project is to get Duet working far enough
so that I can call .send() on torch tensors and send them to
a duet server, and then call .get() on the resulting pointer
and get the tensor back.

So to get started, I'm working in the examples/experimental/trask/Duet Experiments.ipynb
notebook. I start by trying to send a tensor to a duet server and
the main complaint is

TypeError: missing a required argument: 'address'

I remember that this is because I forgot to add address to the
new protobuf file I created for this SaveObjectMessage. And
now in order to do that I need to serialize Address.

but in order to serialize Address, I need to first serialize
SpecificLocation because the address I'm trying to use here
contains a series of SpecificLocation objects.

So that's where my project starts today. Serializing
SpecificLocation as mentioned in the PR I'm working on.'

* Init SpecificLocation serde file

I'm also playing around with SpecificLocation object in the notebook
a little bit. Familiarizing myself with it. I think that I'll improve
the __repr__ a little while I'm here.

* Improve SpecificLocation.__repr__ by improving ObjectWithId.__repr__

The specific location REPR was returnning ObjectWithID by default because
it inherits from it. However, this was a bit misleading. At first I thought
I'd just overload the method in SpecificLocation but I found a better way
which will make all subclasses from ObjectWithID print a more informative
.__repr__() so I changed that instead.

I also had to fix a test which broke because of this improvement.

* Add import to location.proto

The build raised a warning that I was using something I forgot to import. This
fixed that

* Add Serializable methods to SpecificLocation

In order to deserialize, we need to overload the Serialiable methods
in SpecificLocation to use the right protobuf. Interestingly, we
probably could have not done this and modified ObjectWithID to
deserialize correctly. However, instead of scope creep I chose to
go ahead and do a custom class for SpecificLocation despite the
fact that serialization actually already worked since SpecificLocation
doesn't have any additional attributes beyond ObjectWithID. Serde
actually worked, deserializatoin just returned an ObjectWithID
object instead of SpecificLocation (the correct subclass).

* Add docstring for get_protobuf_schema() in all classes that have it

I noticed that SpecificLocation was missing a docstring for
get_protobuf_schema(). Looking around, all classes with this
method seemed to also be missing it. So I added it while
I'm here doing it for SpecificLocation.

* Update get_protobuf_schema() docstring with more precise type

I also found that the Serializable class had a shorter docstring. Fixed that
one too

* Just spotted a documentation problem; Fixed

* Add proto objects to github

I had forgotten to say 'git add .' earlier. Apologies

* Create test module for SpecificLocation +1 test

* Create second test for SpecificLocation init

also add comment to old test which was missing one

* Copy two tests from ObjectWithID -> SpecificLocation

We want to copy in these tests to make sure that they
always work from this calss. It's possible that someday
someone might change the SpecificLocation hierarchy
to no longer subclass from ObjectWithID, but we
need to ensure that these methods still work correctly.

* Init SpecificLocation SERDE Tests

* Explore address in notebook

* Improve broken Address().__repr__()

It was printing strangely as an artifcat of the previous address refactoring

I've also noticed that when address has lots of UIDs that it becomes
really long and difficult to read. We need a short representation
of the UID for when its embedded. I think i'll add that while I'm here.

* Improve Address repr by creating short UID repr

* Init Address protobuf

* Subclass Address from Serializable; Organize imports

I also had to replace UID references with Optional[Location]

* Change UID to Location in Node and Client Subclasses - then refactor

This ended up being a way bigger commit than I expected. Once I started pulling
the thread of tightening Node and Client subclasses with SpecificLocation objects
I needed to tighten up these classes.

- organize the imports into groups
- change UID return statements to Optional[Location] to match parent class (surprised this didn't error before)
- catch a few bugs

* Protobuf was missing

* Fixed issues with non optional SpecificLocation on Client subclasses

* Init address Serializable methods

Because Proto3 doesn't support any optional attributes yet and
address objects have any permutation of location attributes,
we had to implement some extra flags to make sure we didn't
try to deserialize an object that didn't exist

* Init 3 tests of the Address object

* Finish serde tests for Address

* Add address object to SaveObjectAction and serde methods

* Add test for SaveObjectAction

* Eliminate the need for pickle to send client metadata

* Test Duet send

* Init GetObjectAction and GetObjectResponseMessage protos

* Test GetObjectAction

* Almost have tensors sending and receiving; .get() gets empty object

* Almost have tensors sending and receiving; .get() gets empty object

Co-authored-by: Madhava Jay <me@madhavajay.com>