Use id_at_location (#4674)
Currently, we use the id of the object as a way to register an object as a remote location. This brings some advantages, but also some downsides. The advantage is that we will record only once if we send that item multiple times.

ptr = x.send(alice)
ptr = x.send(alice)

Since we send the same object, we keep only 1 instance of that object on alice.

The problem comes (in the above example) when we delete ptr - the second assignment will generate a call to the __del__ method from PointerTensor (this call can come at any time).
If we have the same id, we might execute the second statement and after that, we delete the first ptr --> in this case we will delete the underlying data, where ptr points to.

With this PR we generate another UUID for each sent item - in this scenario, we know that when the first ptr will get deleted it will delete an object that is identified with a UUID that is different than the second one.