feat: create user model
Uses pydantic.

Adds two TODO's as SyftUser expects the role name when creating a user.
However, it is more beneficial if we used the id for role instead of a
string. A new variable role_name could be created for making it easier
to identify the user role for the API consumers.