Add build-essential to dependencies in order to fix building error
The library bplib fails on building when docker image is trying to be created. This cause pip tries to install other versions of bplib and then tries to get all other versions of the other dependencies, taking hours to try to build the image.

Adding build-essential allows to build bplib and then the image is created in a few minutes.