Correcting minor bugs in osx tutorial
Proposing some changes:
Conda environment is set up as `syft_env` in the beginning but referenced as `syft` later. So adding the `_env` where it is missing.
Maybe I missed something, but `pip install syft` is not mentioned in the tutorial. I have added it as step 5.5.

Also, the tutorial is overly specific for MacOs11_5_1. But I have completed it successfully on macOs Monterey 12.1. Maybe consider changing the requirements?